refactor: dao를 repository 인터페이스로 변경
