+++ src/main/java/kitchenpos/domain/repository/MenuGroupRepository.java
@@ -1,11 +1,12 @@
-package kitchenpos.dao;
+package kitchenpos.domain.repository;
 
 import kitchenpos.domain.MenuGroup;
 
 import java.util.List;
 import java.util.Optional;
 
-public interface MenuGroupDao {
+public interface MenuGroupRepository {
+
     MenuGroup save(MenuGroup entity);
 
     Optional<MenuGroup> findById(Long id);
