+++ src/main/java/kitchenpos/domain/repository/MenuProductRepository.java
@@ -1,11 +1,12 @@
-package kitchenpos.dao;
+package kitchenpos.domain.repository;
 
 import kitchenpos.domain.MenuProduct;
 
 import java.util.List;
 import java.util.Optional;
 
-public interface MenuProductDao {
+public interface MenuProductRepository {
+
     MenuProduct save(MenuProduct entity);
 
     Optional<MenuProduct> findById(Long id);
