+++ src/main/java/kitchenpos/domain/repository/OrderTableRepository.java
@@ -1,11 +1,12 @@
-package kitchenpos.dao;
+package kitchenpos.domain.repository;
 
 import kitchenpos.domain.OrderTable;
 
 import java.util.List;
 import java.util.Optional;
 
-public interface OrderTableDao {
+public interface OrderTableRepository {
+
     OrderTable save(OrderTable entity);
 
     Optional<OrderTable> findById(Long id);
