+++ src/main/java/kitchenpos/domain/repository/ProductRepository.java
@@ -1,11 +1,12 @@
-package kitchenpos.dao;
+package kitchenpos.domain.repository;
 
 import kitchenpos.domain.Product;
 
 import java.util.List;
 import java.util.Optional;
 
-public interface ProductDao {
+public interface ProductRepository {
+
     Product save(Product entity);
 
     Optional<Product> findById(Long id);
