+++ src/main/java/kitchenpos/domain/repository/TableGroupRepository.java
@@ -1,11 +1,12 @@
-package kitchenpos.dao;
+package kitchenpos.domain.repository;
 
 import kitchenpos.domain.TableGroup;
 
 import java.util.List;
 import java.util.Optional;
 
-public interface TableGroupDao {
+public interface TableGroupRepository {
+
     TableGroup save(TableGroup entity);
 
     Optional<TableGroup> findById(Long id);
